Forgotten Network size

+++ src/types.js
@@ -27,8 +27,8 @@ var Network = typeforce.compile({
     public: typeforce.UInt32,
     private: typeforce.UInt32
   },
-  pubKeyHash: typeforce.UInt8,
-  scriptHash: typeforce.UInt8,
+  pubKeyHash: typeforce.UInt16,
+  scriptHash: typeforce.UInt16,
   wif: typeforce.UInt8
 })
